+++ app/src/main/java/holamundo/com/tenderos/ClienteMain.java
@@ -76,9 +76,9 @@ public class ClienteMain extends AppCompatActivity implements AdaptadorProducto.
         b.putString("nombre",p.getNombre());
         b.putString("tipo",p.getTipo());
         b.putString("medida",p.getUnidadDeMedida());
-        b.putInt("cantidad",p.getCantidadDisponible());
+        b.putDouble("cantidad",p.getCantidadDisponible());
         b.putDouble("precio",p.getPrecio());
-        b.putInt("foto",p.getFoto());
+        b.putLong("foto",p.getFoto());
 
         i.putExtra("datos",b);
         startActivity(i);
