+++ app/src/main/java/holamundo/com/tenderos/ListaProductos.java
@@ -81,8 +81,8 @@ public class ListaProductos extends AppCompatActivity implements AdaptadorProduc
         b.putString("nombre",p.getNombre());
         b.putString("tipo",p.getTipo());
         b.putString("medida",p.getUnidadDeMedida());
-        b.putInt("cantidad",p.getCantidadDisponible());
-        b.putDouble("precio",p.getPrecio());
+        b.putDouble("cantidad",p.getCantidadDisponible());
+        b.putLong("precio",p.getPrecio());
         b.putInt("foto",p.getFoto());
 
         i.putExtra("datos",b);
